* last_chance_handler: Rename to Simple_Last_Chance_Handler so as not to interract with the  AUnit package.
* support_build.gpr:
Merge with support.gpr to simplify the update of those project files.

* support.gpr:
Merge with support_build.gpr, add handling of AUnit by not including
 the last_chance_handler in this case.

* Makefile.common:
Refine the use of "stty raw". In particular, do not use this command
 with AUnit tests.

* Makefile.examples:
Add support for AUnit tests.

* Makefile:
Compile two versions of the support lib, to handle AUnit tests.

Part of I224-017


git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@1907 020d506d-db78-4e55-b2a8-6c851f84c332