For I527-034 [Ada] Initial complete information for SCO input/output This checkin completes the initial implementation of the circuits for generating SCO (source coverage obligation) information in the ALI file, and also the Get_SCOs routine that reads SCO information back into the internal tables (defined in sco.ads) for use by the coverage tools.
Part of the checkin is a stand alone utility SCO_Test that is used to
test that Get_SCO and Put_SCO are consistent. The way it works is to
compile any file with the -gnateS switch which outputs the coverage
information in the ALI file. Then run

   sco_test file.ali

This causes the SCO information to be read from the ALI file using Get_SCOs and
stored in the internal tables defined in sco.ads. Then Put_SCOs is called to
translate the binary information back to the the text format. Then this output
text is compared to the text in the ALI file to make sure that it is identical.
For example, this can be run on sco_test itself using:

   sco_test sco_test.ali

And the resulting output will be:

  Success, data output by Put_SCOs matches data read by Get_SCOs

This test has been run on an extensive set of ALI files.


-- from svn+ssh://svn.us.adacore.com/Dev/trunk/gnat@150040

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2101 020d506d-db78-4e55-b2a8-6c851f84c332