* par_sco.adb: (Is_Logical_Operator): Exclude AND/OR/XOR
* scos.ads:
Clarify handling of logical operators

For IA01-030
[Ada] SCO output ignores AND/OR/XOR operators
Logical operators for SCO now include only AND THEN/OR ELSE/NOT. This is
in accordance with the latest design, which assumes that we always have
restriction No_Direct_Boolean_Operators when generating SCO's and we are
interested in decision coverage. If we are doing only simple coverage
then this has no effect anyway.

If the following is compiled with -gnateS then grepping
the resulting ali file for "&" should generate no output.

procedure ModularSCO is
   type r is mod 2 ** 8;
   a, b, c : r;
begin
   if (a and b) = c then
      null;
   end if;
end;

And note that the AND here is not a decision in any case, but
just an operation on modular integers. Previous to this patch
this would have been incorrectly treated as a decision.



-- from svn+ssh://svn.us.adacore.com/Dev/trunk/gnat@154923

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2281 020d506d-db78-4e55-b2a8-6c851f84c332