* get_scos.adb: (Get_SCOs): Handle Cs continuation lines
* put_scos.adb:
Implement Cs continuation lines

* scos.ads:
Document use of Cs continuation lines

FOR IA26-040
[Ada] Implement Cs continuation lines for SCO output
Previously, a long basic block could generate a single very long CS
line in the ALI file, which could cause problems. This patch limits
the number of entries on a line to 6, and uses Cs as a continuation
if there are more than 6 items on a line.

The following program, compiled with -gnateS:

procedure SCOlong (X : in out Integer) is
begin
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
   X := X + 1;
end SCOlong;

Yields the following CS/Cs lines in the ali file:

CS 3:4-3:13 4:4-4:13 5:4-5:13 6:4-6:13 7:4-7:13 8:4-8:13
Cs 9:4-9:13 10:4-10:13 11:4-11:13 12:4-12:13 13:4-13:13 14:4-14:13
Cs 15:4-15:13 16:4-16:13 17:4-17:13 18:4-18:13 19:4-19:13 20:4-20:13
Cs 21:4-21:13 22:4-22:13 23:4-23:13 24:4-24:13 25:4-25:13 26:4-26:13
Cs 27:4-27:13 28:4-28:13



-- from svn+ssh://svn.us.adacore.com/Dev/trunk/gnat@155008 by dewar

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2314 020d506d-db78-4e55-b2a8-6c851f84c332