Formal definitions of MC/DC
Give the formal definitions of Unique Cause and MC/DC as defined in
DOT/FAA/AR-01/18, Chilenski, "An Investigation of Three Forms of the
Modified Condition/Decision Coverage (MCDC) Criterion". These formal
definitions will be used as the basis of a "weak" mc/dc criterion that
is needed to prove our theorems about OBC and MC/DC; so it is
worth detailing them a bit here.

Part of IC10-019.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2513 020d506d-db78-4e55-b2a8-6c851f84c332