Model for graph-coloring definition of MC/DC
Add a model for the graph-coloring definition of MC/DC. This will
allow us to check some alternative definitions for Masking MC/DC,
based on paths though BDD rather than relying on the decision syntax
tree.  As xcov makes a extensive usage of BDDs, this could be the
appropriate characterization.

Part of J415-020.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2519 020d506d-db78-4e55-b2a8-6c851f84c332