Fallthrough after delay slot
On certain architectures, the instruction that immediately follows a
branch instruction is executed before the transfer of control; it is
the so-called "delay instruction". In such architecture, the
appropriate address for fallthrough is the pc after this delay
instruction. That helps xcov to identify BRANCH1's destination with
BRANCH2's fallthrough in the following case:

           test
@BRANCH1:  cond branch to address @OUTCOME1
           [...]
           test
@BRANCH2:  cond branch to address @OUTCOME2
           delay instruction
@OUTCOME1: [...]

Fixes J723-011.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2529 020d506d-db78-4e55-b2a8-6c851f84c332