New definition of Masking MC/DC
Express Masking MC/DC in terms of paths through BDD; this definition
would be more easy to implement in xcov. Check that it is equivalent
to the definition given in DOT/FAA/AR-01/18 and check that the
properties that we proved with unique cause regarding branch coverage
are still valid.

Part of J415-020.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2532 020d506d-db78-4e55-b2a8-6c851f84c332