Fix definition of Masking MC/DC
The previous definition of Masking MC/DC in terms of paths through BDD
was bogus, but the error has not been detected because the assertion
constraints (up to 3 conditions) were too strong to generate a
counter-example.  Re-define a correct one and check it with 4
conditions. Rename the previous definition to "breaking_paths" and
prove/document that it is distinct from Masking MC/DC.

Part of J415-020.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2601 020d506d-db78-4e55-b2a8-6c851f84c332