Distinguish short-circuit masking and Masking MC/DC
Short-circuit masking is actually a *phenomenon* that exists with any
definition of MC/DC, as long as short-circuit operators are used: when
considering a evaluation pair, conditions that are not evaluated in
(at least) one of the two evaluations are not considered when proving
an independent influence. On the other hand, Masking MC/DC is a
*coverage criterion* defined for any form of operator (short-circuit
or not). So these are two distinct concepts; e.g. one may observe
short-circuit masking with Unique Cause MC/DC.

This change aims at clarifying this distinction in the testsuite. The
tests that were suffixed with "mask" are actually testing
short-circuit masking, not Masking MC/DC. So rename their suffix as
"sc".  A forthcoming change will add some tests that are specifically
targetted to Masking MC/DC; those will be suffixed with "mask".

Part of JB12-009.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2686 020d506d-db78-4e55-b2a8-6c851f84c332