Do not pass --qualif-cargs to test if value is empty string.
If the testsuite driver is called with --qualif-cargs='', then
it's useless (but not harmful) to pass --qualif-cargs='' to each
test.  So, we don't pass it anymore.

This situation may arise in the context of shell scripts where the testsuite
driver is called like this:

        ./testsuite.py --qualif-cargs="$cargs"

where variable cargs might be null.  Ideally, the script should not
even pass --qualif-cargs is cargs is null, but the above is just
so much easier to write...

Done in the context of J929-016.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2740 020d506d-db78-4e55-b2a8-6c851f84c332