Use Env().discriminants to compute our own list of discriminants
Currently, the list of discriminants that we use in this testsuite
is rather limited (either "ALL", or the platform name (Eg: "ppc-elf").
This allows our test.opt files to test for other features such as
endinaness, CPU name, etc.

This is a small cleanup while working on providing an even larger
pallette of discriminants (allowing us to test for qualif-cargs...).

For J929-016.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2760 020d506d-db78-4e55-b2a8-6c851f84c332