Extends expectation language to distinguish between Masking and Unique Cause
This adds a first support of Masking MC/DC in the testsuite: a way to
separate messages that occur only in Unique Cause ("u!") from
messages that show up in both Unique Cause and Masking ("m!"). Make this
distinction in test cases. No support for this distinction in the driver
(yet) though.

Part of JB12-009

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2775 020d506d-db78-4e55-b2a8-6c851f84c332