Add support for Masking MC/DC in testsuite driver
The tests in directory mcdc are now run for both stmt+mcdc and
stmt+uc_mcdc; expectation notes "u!" are ignored in the case of
Masking MC/DC, and line states ("l!") are fixed in the case when only
"u!" are expected. Note that this resolution strategy for line states
has several drawbacks, as documented in __resolve_lnote.

Part of JB12-009

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2787 020d506d-db78-4e55-b2a8-6c851f84c332