Line states that are dependent on the coverage level
As noted before in __resolve_lnote, the automatic promotion of "l!"
to "l+" in Masking MC/DC presented several problems; in particular
this strategy did not allow to express a different expectation
for a line if no messages were attached to this line in Unique
Cause and Masking MC/DC. This was a problem in
Qualif/Ada/mcdc/OperandS/DiamondS/cc4/src/test_cc4_all.adb
and test_cc4_masking.adb, where a decision (that is 'l+' in
Masking MC/DC, 'l!' in Unique Cause) spans multiple lines.

In order to fix this issue, the expectation language is improved
to express this distinction. Remove __resolve_lnote, now obsolete.

Part of JB12-009

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2801 020d506d-db78-4e55-b2a8-6c851f84c332