Homogeneous treatment of level-dependent expectations
Line expectation notes and report expectation notes used two different
scheme for specifying a dependency on the coverage level:
* for line xnotes, a general scheme (i.e. a way to express things like
"if level=stmt+uc_mcdc, then we expect this, otherwise that");
* for report xnotes, just a set of context-specific notes whose
meaning depends on the coverage level.
This change removes these context-specific notes to use the general scheme
instead. This allows to restore an old invariant: one message of xcov
corresponds exactly to one kind of report note.

Part of JB12-009

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@2805 020d506d-db78-4e55-b2a8-6c851f84c332