Promote Masking MC/DC characterization
The simple characterization of Masking MC/DC in the case of short-circuits
is a nice result that was a bit hidden in the middle of the section about
coupling, as a lemma. Move it to the section of MC/DC properties and
promote it to theorem.

Minor reorg done in the K120-023

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3060 020d506d-db78-4e55-b2a8-6c851f84c332