* put_scos.adb: (Put_SCOs): Do not emit decision SCO for an X decision nested in a disabled pragma.
* scos.ads, scos.adb, par_sco.ads, par_sco.adb:
Record sloc of enclosing pragma, if any, for X decisions.

Fix for K224-017

[Ada] SCO for nested decision in pragma

The decision SCO for the boolean expression in a pragma Assert is emitted
only if assertion checking is enabled. This change ensures that this also
applies to any nested decision within that expression.

The following compilation must produce two decision SCOs (a CP and a CX)
when compiled with -gnata, and none without:

$ gcc -c -gnateS -gnata decision_in_assert.adb
$ grep "^C[PX]" decision_in_assert.ali
CP 6:4 c6:19-6:34
CX &6:25 c6:23-6:23 c6:34-6:34

$ gcc -c -gnateS decision_in_assert.adb
$ grep "^C[PX]" decision_in_assert.ali
<no output>

procedure Decision_In_Assert (A, B : Boolean) is
   function Id (X : Boolean) return Boolean is
   begin
      return X;
   end Id;
   pragma Assert (Id (A and then B));
begin
   null;
end Decision_In_Assert;

-- from svn+ssh://svn.us.adacore.com/Dev/trunk/gnat@173594 by quinot

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3068 020d506d-db78-4e55-b2a8-6c851f84c332