Split SCOV/internals/notes.py into cnotes and xnotep, to separate the note instances from the note patterns, which isolates the pieces that depend on the current test instance and couldn't be referenced from the toplevel suite driver context. Adjust current users accordingly.
Introduce a Qdata module, to serve as a comm medium between testcases
and the suite driver for the production of a test-results qualification
report. Setup the use of it from both sides. No actual report produced
yet. Simplify a few existing constructs alongs the way.

Introduce a "clear" service in SUITE.cutils, and use shorter with/as
constructs throughout.

Part of K407-019



git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3111 020d506d-db78-4e55-b2a8-6c851f84c332