Support sloc with empty address ranges
Such slocs can be generated when some code is optimized out. This
change makes sure that they are not dropped when reading debug
info. As a result, several sloc may be associated to the same PC: so
we either iterate on the list or get the only sloc with a non-empty
range, depending on the situation.

Part of K103-010

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3115 020d506d-db78-4e55-b2a8-6c851f84c332