Add tentative support for couverture tests written in C.
This patch expands the template project file to support C and Ada
sources.  The languages are determined based on the "mains" found
in the source directory. Just like Ada mains, C mains are detected
if the filename starts with "test_" (and ends with the standard ".c"
extension).

The rest of the changes are detection of the C mains, and adjustments
to the generated project files to provide the languages as well.

For IC06-008.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3137 020d506d-db78-4e55-b2a8-6c851f84c332