Store driver filenames rather than exe name in SCOV_helper
We need to store the filename, rather than the executable name,
because we need to remember the language used by this driver.
This also fixes a call to gprbuild where we're forced to add
the ".adb" extension back to pass the filename, which has been
working up to now, but no longer does as soon as drivers can be
written in C as well.

Part of IC06-008.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3185 020d506d-db78-4e55-b2a8-6c851f84c332