Add language-specific processing
This patch introduces a new class LangInfo which provides a description
of some of the aspects that are specific to a given language (filename
extension, comment marker, etc).  We then use that class to guess
the language for a given filename, based on its extension, and then
provide language-specific processing in a generic way.

For IC06-008

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3187 020d506d-db78-4e55-b2a8-6c851f84c332