GNATcoverage qualification material generator - H407-002
- Remove syntactic mistakes introduced in previous change.

- Refine attempt at the production of a general requirements
index, with a separate rst source and a ..index directive.

- Fix thinko in the dirtree reverse-pruning computation, to
account for char-grained (and not dirname-grained) semantics
of os.path.commonprefix. Fixes bogus inclusion of "Qualif/C"
in the material as soon as "Qualif/Common" is included.

- Rename the dirCut node walk control into dirCutPost, to make
it explicit that the walk is cut only past the local node
processing. Introduce dirCutPre, which is useful as well.



git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3433 020d506d-db78-4e55-b2a8-6c851f84c332