(Label_From_Opposite): Do not record a CFG edge valuation deduced by taking the opposite value of the other edge of the same cond branch if that value determines an outcome, and there's already at least one known destination for that outcome, and the destination of the edge does not match any of these known destinations.
This makes the static analysis heuristics more conservative, in order
to gracefully handle the case where a CFG edge for a locally handled
exception is mistaken as an outcome edge.

In this case a warning is generated (on stdout, and in the optional
non-qualified section of the report, if enabled explicitly on the command
line).

Part of K804-031


git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3538 020d506d-db78-4e55-b2a8-6c851f84c332