* exp_intr.adb: (Expand_Unc_Deallocation): Remove test for empty storage pool (this  test is moved to Sem_Intr).
* features-ada2012:
J701-212 AI-0157 Deallocation from empty storage pool

* sem_intr.adb:
(Check_Intrinsic_Call): Add check for deallocation from empty storage
 pool, moved here from Exp_Intr and made into error.
(Check_Intrinsic_Call): Remove assumption in generating not-null free
 warning that the name of the instantiation is Free.

* sinput.adb:
(Tree_Read): Document use of illegal free call allowed in GNAT mode

* types.ads:
Remove storage size clauses from big types (since we may need to
 do deallocations, which are now illegal for empty pools).

For J701-212
For J712-001
[Ada] AI-0157 Deallocation from empty storage pool
This patch makes an attempt to deallocate from an empty storage pool
an error (previously we only gave a warning). The issuing of the message
is moved to the semantics phase so it happens even in -gnatc mode. An odd
glitch is that there is such a call (quite valid) in Sinput, so for now we
fix this by simply not generating the error if we are in GNAT mode.

The following is compiled with -gnatc -gnatld7

     1. with Unchecked_Deallocation;
     2. procedure IllegalUC is
     3.    type R is access Integer;
     4.    for R'Storage_Size use 0;
     5.
     6.    procedure Free is new
     7.      Unchecked_Deallocation (Integer, R);
     8.
     9.    X : R;
    10. begin
    11.    Free (X);
           |
        >>> deallocation from empty storage pool

    12. end;

As a side effect, this patch also fixes a bug in the issuing of the
warning for freeing a not null pointer. This warning assumed that the
name of the instantiation was Free, this assumption is no longer made
as shown by the following test:

     1. with Unchecked_Deallocation;
     2. procedure NNTest is
     3.    type Ptr is access Integer;
     4.    procedure Free_Ptr is new Unchecked_Deallocation (Integer, Ptr);
     5.    Obj : not null Ptr := new Integer;
     6.    X : Ptr := Obj;
     7. begin
     8.    Free_Ptr (Obj);
           |
        >>> warning: freeing "not null" object will raise Constraint_Error

     9.    if Obj = null then
    10.       raise Program_Error;
    11.    elsif Obj /= X then
    12.       raise Program_Error;
    13.    else
    14.       raise Program_Error;
    15.    end if;
    16. end NNTest;

This warning was not previously given




-- from svn+ssh://svn.us.adacore.com/Dev/trunk/gnat@165810 by dewar

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3625 020d506d-db78-4e55-b2a8-6c851f84c332