* s-utf_32.adb: (Is_UTF_32_Line_Terminator): Recognize NEL as line terminator
* sinput.ads:
Add section on Handling of Source Line Terminators

* types.ads:
(Line_Terminator): Adjust comments


[Ada] Properly recognize NEL as end of line in UTF-8 encoding mode

For JB24-034

The NEL (NEXT LINE) control code (16#85#) is now properly recognized as
a line terminator, and hence can be used to terminate a source line when
the source program is encoded in UTF-8 mode.

Note that NEL is not recognized as a line terminator in normal 8-bit
mode. For an extensive explanation of why this is, see the section in
the Sinput spec on "Handling of Source Line Terminators".

The first test program here is to be compiled in normal mode with
no wide character switch, so it is normal Latin-1, in which the
code NEL is not recognized as ending a line:

     1. with Ada.Text_IO; use Ada.Text_IO;
     2. procedure Test_NEL is
     3. begin
     4.    -- Ends with VT so should print?   Put_Line("VT works");
     5.    -- Ends with FF so should print?   Put_Line("FF works");
     6.    -- Ends with NEL so should print?  Put_Line("NEL works");
     7. end Test_NEL;

In the above program
   ? in line 4 is the code 16#0B#, VT
   ? in line 5 is the code 16#0C#, FF
   ? in line 6 is the code 16#85#, NEL

Since NEL is not recognized as ending a line in this mode, the output
of this program is two lines:

VT works
FF works

The second program is to be compiled with the -gnatW8 mode specifying
that it is encoded in UTF-8, where the NEL code is recognized.

     1. with Ada.Text_IO; use Ada.Text_IO;
     2. procedure Test_NEL2 is
     3. begin
     4.    -- Ends with VT so should print?   Put_Line("VT works");
     5.    -- Ends with FF so should print?   Put_Line("FF works");
     6.    -- Ends with NEL so should print?? Put_Line("NEL works");
     7. end Test_NEL2;

In the above program
   ? in line 4 is the code 16#0B#, VT
   ? in line 5 is the code 16#0C#, FF
   ?? in line 6 is the sequence 16#C2# 16#85# (UTF-8 for 16#85#/NEL)

Since NEL is recognized in UTF-8 mode, the output is three lines:

VT works
FF works
NEL works






-- from svn+ssh://svn.us.adacore.com/Dev/trunk/gnat@180313 by dewar

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3766 020d506d-db78-4e55-b2a8-6c851f84c332