(Label_Destination): Generate an arbitrary valuation only for the last edge of a CBI both of whose edges determine a decision outcome. Otherwise we may set a degraded mode origin (i.e. an arbitrary valuation) on the first edge when the second edge can be reliably labeled using dominance information. (If there is no dominance information, the labeling on the first edge will be set during the Label_From_Opposite pass.)
(Trace_Destination): Fix minor error in checking dominance information
(the enclosing decision of a condition is not its parent in the general case
-- the parent is the enclosing operator SCO).

(Load_SCOs): Fix setting of dominance information. Dominance marker slocs
can't be looked up in the general Sloc_To_SCO map while loading SCOs because
the map is constructed later (and furthermore it does not include decisions).
So build a separate D_Sloc_To_SCO map for decisions, and add a separate
pass to set dominance pointers once all SCOs have been loaded.

These changes allow dominance information to be taken into account to provide
more precise information on what branches of a decision are exercised.
This is a collateral benefit of new information generated for K804-031.


git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@3791 020d506d-db78-4e55-b2a8-6c851f84c332