* support/src/libsupport.ad?: Provide the low level mem related 	simple services always (memcmp, memcpy, memmove)
The code for these is simple and target independant here, so not
a maintenance hazard. Having them here makes sure our testing relies
on the same code for every configuration, which actually is a nice
property. Besides, this allows support coverage tests for a wide range
of constructs that really don't commit RTS violation per se at the
user level, unlike one using an explicit allocator for example.

Part of L814-007.



git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@4075 020d506d-db78-4e55-b2a8-6c851f84c332