Step on the productization of gnatcov for x86-linux. LB16-016
The general purpose of this change is to let gnatcov locate a local
version of valgrind and to let it update the VALGRIND_LIB environment
variable prior to execution, so whatever valgrind executes knows where
to find our coverage plugin. This way, we can have our plugin installed
locally, even if we end up using a user-provided valgrind eventually.

The idea is to

- Generalize our internal macro expansion circuitry so it also
  applies to commands in our configuration database, not only to
  argument lists,

- Rename Build_Command into Setup_Command in our configuration
  tables, more representative of the current intent.

- Move the macro expansion code in a private child unit, which
  makes the whole, growing, thing easier to grasp.

- Add comments ...

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@4235 020d506d-db78-4e55-b2a8-6c851f84c332