Further refinements to the valgrind bundling in gnatcov. LB16-016, LB26-006
Remove extraneous quoting on Makefile variable assignment.

Adjust gnatcov to search for valgrind in libexec/bin and
to not set VALGRIND_LIB when resorting to the valgrind on PATH.

VALGRIND_LIB is used to find all the valgrind modules, including core 
ones, and we don't want to have our local set of libs (in libexec) used 
with a possibly unrelated version of valgrind (when taken from PATH).

Our coverage plugin must be installed within the valgrind on PATH tree
in this case as well.



git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@4256 020d506d-db78-4e55-b2a8-6c851f84c332