testsuite: create a dedicated HarnessTestCase class for expected errors
M304-036

Harness test cases exist to verify that the testsuite detects mismatches
between test expectations and actual results. For such tests, the testsuite
framework must not raise errors when it is expected that some tool (xcov,
xrun, ...) fails. The HarnessTestCase class handles this thanks to
coordination with the SCOV_helper class.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@4393 020d506d-db78-4e55-b2a8-6c851f84c332