testsuite: add SCOV.expgen: testcases generation for LibExp
Ongoing work on L911-012

The testcases generation rely on test drivers: decision expression
topologies are read from them, and controllers for each language, operand
kinds and decision contexts.

Operand kinds and contexts can be based both on language constructs that
are common to C and Ada, or on language-specific constructs.

Generator in-source documentation (in the SCOV.expgen package) will be
completed when more tests will be added to demonstrate the validity of this
approach.

git-svn-id: svn+ssh://scm.forge.open-do.org/scmrepos/svn/couverture/trunk/couverture@4412 020d506d-db78-4e55-b2a8-6c851f84c332