Adjust stmt/Robustness tests not to use direct boolean operators (M421-004)
Allowing the tests to pass with Restrictions (No_Direct_Boolean_Operators),
about to be enforced for level A qualification runs.

The use of direct operators was not relevant the tests at hand so
the change is correct wrt original test intent (checking the tool
behavior on multiline statements and multi statements lines).