Preparatory work for M420-001 (building qgnatcov qualkit documents)
Distinguish KNOWN_LANGUAGES, languages that gnatcov supports
and which the testsuite needs to know about from QLANGUAGES, languages
for which we might be generating qualification material.

QLANGUAGES used to be used for both purposes, creating confusion
since we improved our support for C and started to populate the Qualif/C
tree even though we're not at all ready for qualification yet.