Improvements to static analysis heuristics
This change ensures that a conditional branch instruction edge that branches
out of its enclosing statement is correctly treated as a decision outcome.
The heuristics for labeling an edge from the opposite one is relaxed to
also work when the edge destination is not previously known, even if
other destinations with the same outcome are already known, as long as
no inconsistency is introduced.

For M422-034