Improve inlined subprograms handling using corresponding DWARF info
M419-017

It lacks support of addresses ranges for inlined subroutines DIEs
(DW_TAG_inlined_subroutine, see the added TODO).

This is not going to be completed however, since the whole commit is useless
as long as the compiler add smartly-sloc'ed NOP instructions to inlined
subroutines. Thanks to those, gnatcov computes automagically the coverage
for call statements that caused inlining.