Accurate handling of multi-statement lines with precise debug info
This change improves the processing for lines with multiple statements:
degraded (partial) coverage is only reported when debug information is
actually insufficient to establish whether a statement has been executed.
If precise debug info (including column numbers) is provided for statements,
full coverage can be achieved.

This change also has a nice side effect that if we propagate from
a subsequent (single-statement) line back to a multi-statement one,
we no longer report partial coverage on the latter, as we know for
certain all of its statements have been executed. (And this is true
even in the absence of column-precise debug info for statements).

For M514-031