scripts/decision_cfg: add --target to let the user specify toolchain prefix
Feature requested in M523-039

The script cannot guess what toolchain to use: for a PowerPC program, the
name of the objdump program to use can be "powerpc-elf-objdump", but also
"powerpc-wrs-vxworks-objdump", or even just "objdump". It is now up to the
user to specify the toolchain prefix, thanks to a new --target option.

If not given, no prefix is prepend to "objdump", and if a target is given,
"TARGET-objdump" is used. In the case no such program is found, an
appropriate error message is displayed:

    decision_cfg.py: error: argument --target: No powerpc-elf-objdump found