Add error and message for bad start trigger choice.
If Start Trigger is at a branch instruction, it causes
miscaluculation of location of a block of executed instructions.
This was alway caught (and caused a generic internal error
report), but this condition is now found earlier and explicitly,
and results in a specific errror message describing the user
change that is needed.

While testing this saw a minor error message formatting error
from gnatcov (in convert.adb).

For M628-014