scripts/intervalmap: fix a bug in the bounds registry
Ongoing work on M424-015

A typo made bounds registering buggy: for special cases of insertion, the
high bound was inserted instead of the low bound, resulting in buggy
internal data. This is now fixed, and a sanity check has been added to unit
tests in odrer to checks for such bugs.