Wording correction
An earlier version of the file said "nothing is reported about
statements in subprograms that are called".  This seemed confusing,
since such statements are covered, so it was changed to
"subprograms that are called are reported as covered".  But this
is not correct, since what is reported are coverage violations.
So the wording was changed again, this time to "subprograms that
are called are not reported as uncovered".
M607-003