Reorganize circuitry for elimination of non-contributive branches
In the presence of transient blocks and exception handlers, the compiler
generates a number of additional conditional branch instructions, which
do not correspond to any source condition, although their slocs fall within
the range of conditions and decisions (because the transient blocks are
introduced by the compiler front-end to account for some properties of
an expression, e.g. the use of intermediate controlled objects or of
the secondary stack). This change contributes to ensuring that these
additional conditional branches, appearing in a decision occurrence after
the outcome has been determined, are excluded from further processing,
since they are irrelevant with respect to source coverage analysis.

Note that compiler changes are also useful in this respect, consisting in
changing the slocs for some of these additional conditional branches so that
they have null column information, and thus are not associated by gnatcov
with any specific source condition.

Part of M724-035