testsuite: fix the C FOR loop related test to remove "no-code" issues
Followup on M319-018

The test controller used to erase some data (the variable "a"). At -O1,
the compiler could then remove the statements that generated this data
("++a"). As a result, the CFG layout could be simplified so that a
"goto" statement emitted no code: this caused trouble with respects to
coverage expectations.

Change-Id: I214afcb26b24f574da6736ae85b14434346cd30b