Introduce the use of gnatpython.mainloop.add_mainloop_options
This function adds standard testsuite options to the driver,
instead of letting the driver implement them all by itself.
For now, the funtion is callled with extended_options=False,
which means that only 2 options are being added: -j/--jobs,
and -abort-file. This patch removes the code adding -j/--jobs
in the driver itself.

This is preparation work for M602-006, where we will want to
introduce --skip-if-* command-line options.