Add support for --skip-if-ok mainloop command-line option.
This patch adds a first version allowing a user to run the testsuite
a second time, re-using the results from the testcases that previously
passed (result == 'OK' or 'UOK'). To enable this feature, use the new
--skip-if-ok command line option.

At this moment, no attempt has been made to verify that subsequent runs
are performed consistently with the first run. For instance, it does
not verify that the same version of the compiler, or gnatcoverage
got used in the two runs. This remains the user's responsibility.

For M602-006.