Rework TOR introductory matters
Move general comments on the document origanization
from the "testsuite description" appendix to the "introduction"
chapter.

Specialize the "testsuite description" appendix as a detailed
overview of the testcase structure + a high level overview of
the testsuite execution harness.

Thicken the toplevel "presentation" a bit. Thicken the various
definitions exposed by the "Introduction" section.

Ongoing work for M930-047