gnatcov genbundle: refine kitname computation, adjust stamp format
Adjust the stamp part of the name to be YYYYMMDD without dashes, as
in the gnatcheck case and matching the qa directory format advertised
in the qualification plans.

Arrange not to duplicate "qualkit" in the kit filename if that happens
to be part of the git branch name already.

Ongoing work for MA18-014