Add a new report format: dynamic HTML.
This format uses two resources files generated as part of the GNATquilt
project. An installer script is provided to install these files in the
GNATcoverage installation.

The new format, dynamic HTML, available with the --annotate=dhtml switch
for gnatcov, is available only if those resources have been installed.
This operation is currently manual but may be automatized as part of the
GNATcov build procedure (via the corresponding Anod file).

It provides a single standalone HTML file containing all the
dependencies inlined.

tools/gnatcov/ChangeLog:

    * annotations.ad[bs] (Annotation_Format,To_Annotate_Format): add new
      format for the dynamic HTML output.

    * gnatcov.adb: Update usage; check for a correct installation before
      generating the output if report format is dynamic HTML.

    * annotations-dynamic_html.ad[bs]: new package for generating the
      new dynamic HTML report.

TN: M405-014