gnatcov PLANS: generalize item on qualification data in conf recommendations
So it encompasses the archives for all the items described in the
qualification data overview section instead of re-listing all of them
individually. This section's purpose is only to offer suggestions, and
whether/how we are doing QA on each item is orthogonal.

In the course of MB14-040, preparing adjustments to come for MA30-051
(purpose of html) and MA30-054 (location of the as-run test results).