Optimize loading time of Dynamic HTML reports.
Loading of big project could be quite long (several seconds) because of
all the data that needs to be loaded in the browser.

This patch enhances the loading time by splitting the JSON structure
into much smaller files (for now the granularity is one hunk file by
source file) and having the HTML report loading them lazily.

The output is now:
    - one index.html file
    - multiple .js files

The HTML file stays "standalone" in the sense that it contains all the
necessary information to display the summary table without any external
dependency file. However, displaying the content of a source file
requires the correct .js file to be available in the same directory.

tools/gnatcov/ChangeLog:

    * annotations-dynamic_html.adb: Generate_Report now generates
      multiple .js files in the output directory along with the main
      index.html.

TN: M405-014