Stream OCD to Nexus to Gnatcov trace info.
From the file containing the On Chip Degub (OCD) info, the
internal Nexus trace data representation was created as one
list before conversion to the trace format used by gnatcov.
This took too much space (and time?) for large traces.

These changes implement the creation of Nexus messages one at a
time as the conversion to gnatcov format is processing.

For : MC06-043