gnatcov suite: preliminary support for pre/post testsuite hooks
Add support for --pre-testcase, --pre-testsuite and --post-testsuite
command line options.

Add support for a --altrun command line option, which allows wrapping up
binaries to use as implicit values for a subset of the --pre|post hooks,
as well as for a subset of the --gnatcov-<cmd> family.

Unify the processing of both families of hooks (declaration of the
support sets, rename the --gnatcov- series as gnatcov_, ...)

Extract path resolution out of parse_options and defer past the
alternate-run hooks setup, which takes place after the first option
parsing pass and influences options contents.

Move the logdir initialization early, before the options and hooks
setup calls as at least the latter dumps an 'altrun' log for debugging
purposes.

Add toplevel documentation of the overall scheme.

Part of N113-041